reduce code smells
