+++ core/src/space/hypeo/networking/endpoint/IEndpoint.java
@@ -1,6 +1,5 @@
 package space.hypeo.networking.endpoint;
 
-import space.hypeo.networking.network.PlayerBusiness;
 import space.hypeo.networking.network.PlayerSkeleton;
 import space.hypeo.networking.packages.Remittances;
 
@@ -13,21 +12,6 @@ public interface IEndpoint {
     public void stop();
 
     /**
-     * Update view of game lobby.
-     */
-    //public void updateStageLobby() {
-        /*
-        Log.info(role + ": updateStageLobby");
-
-        Stage currentStage = stageManager.getCurrentStage();
-
-        if (currentStage instanceof LobbyStage) {
-            ((LobbyStage) currentStage).updateLobby();
-        }
-        */
-    //}
-
-    /**
      * Toggles the status of a player in the lobby between:
      * ready2play <-> not_ready2play
      *
@@ -37,7 +21,6 @@ public interface IEndpoint {
 
     /**
      * Resends a received MoneyAmount from player to another player.
-     *
      * @param remittances
      */
     public void changeBalance(Remittances remittances);
